websocket ready state after order book initialization. refactor. Proper break-even price. Binance server time.

+++ src/main/java/org/tradebot/domain/Imbalance.java
@@ -76,7 +76,7 @@ public class Imbalance implements Serializable {
     }
 
     public double speed() {
-        return size() / duration();
+        return size() / (double) duration();
     }
 
     @Override
